zadania/Z0304/requirements.txt
Flask
pandas
python-dotenv
requests